[cleanup] plugin desc change and a .build.cs cleanup

+++ Source/ProjectCleaner/Private/Commandlets/ProjectCleanerCLICommandlet.cpp
@@ -50,7 +50,10 @@ int32 UProjectCleanerCLICommandlet::Main(const FString& Params)
 		}
 		
 		UE_LOG(LogProjectCleanerCLI, Display, TEXT("Deleted: %d assets"), CleanerDataManager.DeleteAllUnusedAssets());
-		UE_LOG(LogProjectCleanerCLI, Display, TEXT("Deleted: %d empty folders"), CleanerDataManager.DeleteEmptyFolders());
+		if (bAutomaticallyDeleteEmptyFolders)
+		{
+			UE_LOG(LogProjectCleanerCLI, Display, TEXT("Deleted: %d empty folders"), CleanerDataManager.DeleteEmptyFolders());
+		}
 		UE_LOG(LogProjectCleanerCLI, Display, TEXT("===================================="));
 		UE_LOG(LogProjectCleanerCLI, Display, TEXT("=========  After Cleanup    ========"));
 		UE_LOG(LogProjectCleanerCLI, Display, TEXT("===================================="));
@@ -81,7 +84,7 @@ void UProjectCleanerCLICommandlet::ParseCommandLinesArguments(const FString& Par
 	// -Check - false
 	// -ScanDevContent - false
 	// -DeleteEmptyFolders - true
-	// -ExcludeAssets - empty // todo:ashe23
+	// -ExcludeAssets - empty
 	// -ExcludeAssetsInPath - empty 
 	// -ExcludeAssetWithClass - empty
 	if (Switches.Num() == 0 && Parameters.Num() == 1 && Tokens.Num() == 0) // Parameters contain -run=ProjectCleanerCLI - argument only
